+++ src/plugins/discover/public/context_awareness/profile_providers/observability/consts.ts
@@ -7,4 +7,4 @@
  * License v3.0 only", or the "Server Side Public License, v 1".
  */
 
-export { createLogDocumentProfileProvider } from './profile';
+export const OBSERVABILITY_ROOT_PROFILE_ID = 'observability-root-profile';
