+++ src/plugins/discover/public/context_awareness/profile_providers/observability/log_document_profile/accessors/index.ts
