+++ src/plugins/discover/public/context_awareness/profile_providers/observability/logs_data_source_profile/create_profile_providers.ts
@@ -19,7 +19,9 @@ import {
   createWindowsLogsDataSourceProfileProvider,
 } from './sub_profiles';
 
-export const createLogsDataSourceProfileProviders = (providerServices: ProfileProviderServices) => {
+export const createObservabilityLogsDataSourceProfileProviders = (
+  providerServices: ProfileProviderServices
+) => {
   const logsDataSourceProfileProvider = createLogsDataSourceProfileProvider(providerServices);
 
   return [
