+++ src/plugins/discover/public/context_awareness/profile_providers/observability/logs_data_source_profile/index.ts
@@ -7,4 +7,4 @@
  * License v3.0 only", or the "Server Side Public License, v 1".
  */
 
-export { createLogsDataSourceProfileProviders } from './create_profile_providers';
+export { createObservabilityLogsDataSourceProfileProviders } from './create_profile_providers';
