+++ src/plugins/discover/public/context_awareness/profile_providers/observability/logs_data_source_profile/sub_profiles/index.ts
