+++ src/plugins/discover/public/context_awareness/profile_providers/observability/logs_data_source_profile/sub_profiles/system_logs.test.ts
@@ -12,21 +12,26 @@ import { createEsqlDataSource } from '../../../../../../common/data_sources';
 import { DataSourceCategory, RootContext, SolutionType } from '../../../../profiles';
 import { createContextAwarenessMocks } from '../../../../__mocks__';
 import { createLogsDataSourceProfileProvider } from '../profile';
-import { createKubernetesContainerLogsDataSourceProfileProvider } from './kubernetes_container_logs';
+import { createSystemLogsDataSourceProfileProvider } from './system_logs';
+import { ContextWithProfileId } from '../../../../profile_service';
+import { OBSERVABILITY_ROOT_PROFILE_ID } from '../../consts';
 
-const ROOT_CONTEXT: RootContext = { solutionType: SolutionType.Default };
+const ROOT_CONTEXT: ContextWithProfileId<RootContext> = {
+  profileId: OBSERVABILITY_ROOT_PROFILE_ID,
+  solutionType: SolutionType.Observability,
+};
 const { profileProviderServices } = createContextAwarenessMocks();
 const logsDataSourceProfileProvider = createLogsDataSourceProfileProvider(profileProviderServices);
-const dataSourceProfileProvider = createKubernetesContainerLogsDataSourceProfileProvider(
+const dataSourceProfileProvider = createSystemLogsDataSourceProfileProvider(
   logsDataSourceProfileProvider
 );
 
-describe('createKubernetesContainerLogsDataSourceProfileProvider', () => {
+describe('createSystemLogsDataSourceProfileProvider', () => {
   it('should match a valid index pattern', async () => {
     const result = await dataSourceProfileProvider.resolve({
       rootContext: ROOT_CONTEXT,
       dataSource: createEsqlDataSource(),
-      query: { esql: 'FROM logs-kubernetes.container_logs-*' },
+      query: { esql: 'FROM logs-system.syslog-*' },
     });
     expect(result).toEqual({ isMatch: true, context: { category: DataSourceCategory.Logs } });
   });
@@ -35,7 +40,7 @@ describe('createKubernetesContainerLogsDataSourceProfileProvider', () => {
     const result = await dataSourceProfileProvider.resolve({
       rootContext: ROOT_CONTEXT,
       dataSource: createEsqlDataSource(),
-      query: { esql: 'FROM logs-kubernetes.access_logs-*' },
+      query: { esql: 'FROM logs-notsystem.syslog-*' },
     });
     expect(result).toEqual({ isMatch: false });
   });
@@ -48,9 +53,8 @@ describe('createKubernetesContainerLogsDataSourceProfileProvider', () => {
       columns: [
         { name: 'timestamp', width: 212 },
         { name: 'log.level', width: 150 },
-        { name: 'kubernetes.pod.name', width: 200 },
-        { name: 'kubernetes.namespace', width: 200 },
-        { name: 'orchestrator.cluster.name', width: 200 },
+        { name: 'process.name', width: 150 },
+        { name: 'host.name', width: 250 },
         { name: 'message' },
       ],
     });
