+++ src/plugins/discover/public/context_awareness/profile_providers/register_profile_providers.ts
@@ -19,8 +19,8 @@ import {
   createExampleSolutionViewRootProfileProvider,
   createExampleRootProfileProvider,
 } from './example/example_root_profile';
-import { createLogsDataSourceProfileProviders } from './common/logs_data_source_profile';
-import { createLogDocumentProfileProvider } from './common/log_document_profile';
+import { createObservabilityLogsDataSourceProfileProviders } from './observability/logs_data_source_profile';
+import { createObservabilityLogDocumentProfileProvider } from './observability/log_document_profile';
 import { createSecurityRootProfileProvider } from './security/security_root_profile';
 import {
   createProfileProviderServices,
@@ -133,7 +133,7 @@ const createRootProfileProviders = (providerServices: ProfileProviderServices) =
  */
 const createDataSourceProfileProviders = (providerServices: ProfileProviderServices) => [
   createExampleDataSourceProfileProvider(),
-  ...createLogsDataSourceProfileProviders(providerServices),
+  ...createObservabilityLogsDataSourceProfileProviders(providerServices),
 ];
 
 /**
@@ -143,5 +143,5 @@ const createDataSourceProfileProviders = (providerServices: ProfileProviderServi
  */
 const createDocumentProfileProviders = (providerServices: ProfileProviderServices) => [
   createExampleDocumentProfileProvider(),
-  createLogDocumentProfileProvider(providerServices),
+  createObservabilityLogDocumentProfileProvider(providerServices),
 ];
