+++ src/plugins/discover/public/context_awareness/profiles/document_profile.ts
@@ -9,7 +9,7 @@
 
 import type { DataTableRecord } from '@kbn/discover-utils';
 import type { Profile } from '../types';
-import { ProfileProvider, ProfileService } from '../profile_service';
+import { ContextWithProfileId, ProfileProvider, ProfileService } from '../profile_service';
 import type { RootContext } from './root_profile';
 import type { DataSourceContext } from './data_source_profile';
 
@@ -33,11 +33,11 @@ export interface DocumentProfileProviderParams {
   /**
    * The current root context
    */
-  rootContext: RootContext;
+  rootContext: ContextWithProfileId<RootContext>;
   /**
    * The current data source context
    */
-  dataSourceContext: DataSourceContext;
+  dataSourceContext: ContextWithProfileId<DataSourceContext>;
   /**
    * The current data table record
    */
