+++ test/functional/apps/discover/context_awareness/extensions/_get_cell_renderers.ts
@@ -9,344 +9,127 @@
 
 import kbnRison from '@kbn/rison';
 import expect from '@kbn/expect';
-import type { WebElementWrapper } from '@kbn/ftr-common-functional-ui-services';
 import type { FtrProviderContext } from '../../ftr_provider_context';
 
 export default function ({ getService, getPageObjects }: FtrProviderContext) {
-  const { common, discover, unifiedFieldList, header } = getPageObjects([
+  const { common, discover, unifiedFieldList } = getPageObjects([
     'common',
     'discover',
     'unifiedFieldList',
-    'header',
   ]);
-  const esArchiver = getService('esArchiver');
   const testSubjects = getService('testSubjects');
-  const dataGrid = getService('dataGrid');
   const dataViews = getService('dataViews');
-  const queryBar = getService('queryBar');
-  const browser = getService('browser');
-  const retry = getService('retry');
 
   describe('extension getCellRenderers', () => {
-    before(async () => {
-      await esArchiver.loadIfNeeded('test/functional/fixtures/es_archiver/logstash_functional');
-    });
-
-    after(async () => {
-      await esArchiver.unload('test/functional/fixtures/es_archiver/logstash_functional');
-    });
-
     describe('ES|QL mode', () => {
-      describe('Log Level Badge Cell', () => {
-        it('should render log.level badge cell', async () => {
+      describe('root profile', () => {
+        it('should render custom @timestamp', async () => {
           const state = kbnRison.encode({
             dataSource: { type: 'esql' },
-            query: {
-              esql: 'from my-example-logs,logstash* | sort @timestamp desc | where `log.level` is not null',
-            },
+            query: { esql: 'from my-example-* | sort @timestamp desc' },
           });
           await common.navigateToActualUrl('discover', `?_a=${state}`, {
             ensureCurrentUrl: false,
           });
-          await header.waitUntilLoadingHasFinished();
           await discover.waitUntilSearchingHasFinished();
-          await unifiedFieldList.clickFieldListItemAdd('log.level');
-          await header.waitUntilLoadingHasFinished();
-          await discover.waitUntilSearchingHasFinished();
-
-          const firstCell = await dataGrid.getCellElementExcludingControlColumns(0, 0);
-          const logLevelBadge = await firstCell.findByTestSubject('*logLevelBadgeCell-');
-          expect(await logLevelBadge.getVisibleText()).to.be('debug');
-          expect(await logLevelBadge.getComputedStyle('background-color')).to.be(
-            'rgba(190, 207, 227, 1)'
-          );
-        });
-
-        it("should not render log.level badge cell if it's not a logs data source", async () => {
-          const state = kbnRison.encode({
-            dataSource: { type: 'esql' },
-            query: {
-              esql: 'from my-example* | sort @timestamp desc | where `log.level` is not null',
-            },
-          });
-          await common.navigateToActualUrl('discover', `?_a=${state}`, {
-            ensureCurrentUrl: false,
-          });
-          await header.waitUntilLoadingHasFinished();
-          await discover.waitUntilSearchingHasFinished();
-          await unifiedFieldList.clickFieldListItemAdd('log.level');
-          await header.waitUntilLoadingHasFinished();
-          await discover.waitUntilSearchingHasFinished();
-
-          await retry.try(async () => {
-            const firstCell = await dataGrid.getCellElementExcludingControlColumns(0, 0);
-            expect(await firstCell.getVisibleText()).to.be('debug');
-            await testSubjects.missingOrFail('*logLevelBadgeCell-');
-          });
+          const timestamps = await testSubjects.findAll('exampleRootProfileTimestamp');
+          expect(timestamps).to.have.length(6);
+          expect(await timestamps[0].getVisibleText()).to.be('2024-06-10T16:30:00.000Z');
+          expect(await timestamps[5].getVisibleText()).to.be('2024-06-10T14:00:00.000Z');
         });
       });
 
-      describe('Service Name Cell', () => {
-        it('should render service.name cell', async () => {
+      describe('data source profile', () => {
+        it('should render custom @timestamp but not custom log.level', async () => {
           const state = kbnRison.encode({
             dataSource: { type: 'esql' },
-            query: {
-              esql: 'from my-example-logs,logstash* | sort @timestamp desc | where `service.name` is not null',
-            },
+            query: { esql: 'from my-example-* | sort @timestamp desc' },
           });
           await common.navigateToActualUrl('discover', `?_a=${state}`, {
             ensureCurrentUrl: false,
           });
-          await header.waitUntilLoadingHasFinished();
           await discover.waitUntilSearchingHasFinished();
-          await unifiedFieldList.clickFieldListItemAdd('service.name');
-          await header.waitUntilLoadingHasFinished();
-          await discover.waitUntilSearchingHasFinished();
-
-          const firstCell = await dataGrid.getCellElementExcludingControlColumns(0, 0);
-          const lastCell = await dataGrid.getCellElementExcludingControlColumns(2, 0);
-          const firstServiceNameCell = await firstCell.findByTestSubject(
-            'dataTableCellActionsPopover_service.name'
-          );
-          const lastServiceNameCell = await lastCell.findByTestSubject(
-            'dataTableCellActionsPopover_service.name'
-          );
-          expect(await firstServiceNameCell.getVisibleText()).to.be('product');
-          expect(await lastServiceNameCell.getVisibleText()).to.be('accounting');
+          await unifiedFieldList.clickFieldListItemAdd('@timestamp');
+          await unifiedFieldList.clickFieldListItemAdd('log.level');
+          const timestamps = await testSubjects.findAll('exampleRootProfileTimestamp');
+          expect(timestamps).to.have.length(6);
+          expect(await timestamps[0].getVisibleText()).to.be('2024-06-10T16:30:00.000Z');
+          expect(await timestamps[5].getVisibleText()).to.be('2024-06-10T14:00:00.000Z');
+          const logLevels = await testSubjects.findAll('exampleDataSourceProfileLogLevel', 2500);
+          expect(logLevels).to.have.length(0);
         });
 
-        it("should not render service.name cell if it's not a logs data source", async () => {
+        it('should render custom @timestamp and custom log.level', async () => {
           const state = kbnRison.encode({
             dataSource: { type: 'esql' },
-            query: {
-              esql: 'from my-example* | sort @timestamp desc | where `service.name` is not null',
-            },
+            query: { esql: 'from my-example-logs | sort @timestamp desc' },
           });
           await common.navigateToActualUrl('discover', `?_a=${state}`, {
             ensureCurrentUrl: false,
           });
-          await header.waitUntilLoadingHasFinished();
           await discover.waitUntilSearchingHasFinished();
-          await unifiedFieldList.clickFieldListItemAdd('service.name');
-          await header.waitUntilLoadingHasFinished();
-          await discover.waitUntilSearchingHasFinished();
-
-          await retry.try(async () => {
-            const firstCell = await dataGrid.getCellElementExcludingControlColumns(0, 0);
-            expect(await firstCell.getVisibleText()).to.be('product');
-            await testSubjects.missingOrFail('dataTableCellActionsPopover_service.name');
-          });
-        });
-      });
-
-      describe('Summary column', () => {
-        it('should render a summary of the log entry replacing the original document', async () => {
-          const state = kbnRison.encode({
-            dataSource: { type: 'esql' },
-            query: {
-              esql: 'from my-example-logs,logstash* | sort @timestamp desc | where `message` is not null',
-            },
-          });
-          await common.navigateToActualUrl('discover', `?_a=${state}`, {
-            ensureCurrentUrl: false,
-          });
-          await header.waitUntilLoadingHasFinished();
-          await discover.waitUntilSearchingHasFinished();
-
-          await testSubjects.existOrFail('discoverDataTableMessageValue');
-        });
-
-        it('should NOT render the summary column if the source does not match logs', async () => {
-          const state = kbnRison.encode({
-            dataSource: { type: 'esql' },
-            query: {
-              esql: 'from my-example-*',
-            },
-          });
-          await common.navigateToActualUrl('discover', `?_a=${state}`, {
-            ensureCurrentUrl: false,
-          });
-          await header.waitUntilLoadingHasFinished();
-          await discover.waitUntilSearchingHasFinished();
-
-          await testSubjects.missingOrFail('discoverDataTableMessageValue');
+          await unifiedFieldList.clickFieldListItemAdd('@timestamp');
+          await unifiedFieldList.clickFieldListItemAdd('log.level');
+          const timestamps = await testSubjects.findAll('exampleRootProfileTimestamp');
+          expect(timestamps).to.have.length(3);
+          expect(await timestamps[0].getVisibleText()).to.be('2024-06-10T16:00:00.000Z');
+          expect(await timestamps[2].getVisibleText()).to.be('2024-06-10T14:00:00.000Z');
+          const logLevels = await testSubjects.findAll('exampleDataSourceProfileLogLevel');
+          expect(logLevels).to.have.length(3);
+          expect(await logLevels[0].getVisibleText()).to.be('Debug');
+          expect(await logLevels[2].getVisibleText()).to.be('Info');
         });
       });
     });
 
     describe('data view mode', () => {
-      describe('Log Level Badge Cell', () => {
-        it('should render log.level badge cell', async () => {
+      describe('root profile', () => {
+        it('should render custom @timestamp', async () => {
           await common.navigateToActualUrl('discover', undefined, {
             ensureCurrentUrl: false,
           });
-          await header.waitUntilLoadingHasFinished();
+          await dataViews.switchTo('my-example-*');
           await discover.waitUntilSearchingHasFinished();
-          await dataViews.switchToAndValidate('my-example-logs,logstash*');
-          await queryBar.setQuery('log.level:*');
-          await queryBar.submitQuery();
-          await header.waitUntilLoadingHasFinished();
-          await discover.waitUntilSearchingHasFinished();
-          await unifiedFieldList.clickFieldListItemAdd('log.level');
-          await header.waitUntilLoadingHasFinished();
-          await discover.waitUntilSearchingHasFinished();
-
-          let firstCell: WebElementWrapper;
-          let logLevelBadge: WebElementWrapper;
-
-          await retry.try(async () => {
-            firstCell = await dataGrid.getCellElementExcludingControlColumns(0, 1);
-            logLevelBadge = await firstCell.findByTestSubject('*logLevelBadgeCell-');
-            expect(await logLevelBadge.getVisibleText()).to.be('debug');
-            expect(await logLevelBadge.getComputedStyle('background-color')).to.be(
-              'rgba(190, 207, 227, 1)'
-            );
-          });
-
-          // check Surrounding docs page
-          await dataGrid.clickRowToggle();
-          const [, surroundingActionEl] = await dataGrid.getRowActions();
-          await surroundingActionEl.click();
-          await header.waitUntilLoadingHasFinished();
-          await browser.refresh();
-          await header.waitUntilLoadingHasFinished();
-
-          await retry.try(async () => {
-            firstCell = await dataGrid.getCellElementExcludingControlColumns(0, 1);
-            logLevelBadge = await firstCell.findByTestSubject('*logLevelBadgeCell-');
-            expect(await logLevelBadge.getVisibleText()).to.be('debug');
-            expect(await logLevelBadge.getComputedStyle('background-color')).to.be(
-              'rgba(190, 207, 227, 1)'
-            );
-          });
-        });
-
-        it("should not render log.level badge cell if it's not a logs data source", async () => {
-          await common.navigateToActualUrl('discover', undefined, {
-            ensureCurrentUrl: false,
-          });
-          await header.waitUntilLoadingHasFinished();
-          await discover.waitUntilSearchingHasFinished();
-          await dataViews.switchToAndValidate('my-example-*');
-          await queryBar.setQuery('log.level:*');
-          await queryBar.submitQuery();
-          await header.waitUntilLoadingHasFinished();
-          await discover.waitUntilSearchingHasFinished();
-          await unifiedFieldList.clickFieldListItemAdd('log.level');
-          await header.waitUntilLoadingHasFinished();
-          await discover.waitUntilSearchingHasFinished();
-
-          let firstCell: WebElementWrapper;
-
-          await retry.try(async () => {
-            firstCell = await dataGrid.getCellElementExcludingControlColumns(0, 1);
-            expect(await firstCell.getVisibleText()).to.be('debug');
-            await testSubjects.missingOrFail('*logLevelBadgeCell-');
-          });
-
-          // check Surrounding docs page
-          await dataGrid.clickRowToggle();
-          const [, surroundingActionEl] = await dataGrid.getRowActions();
-          await surroundingActionEl.click();
-          await header.waitUntilLoadingHasFinished();
-          await browser.refresh();
-          await header.waitUntilLoadingHasFinished();
-
-          await retry.try(async () => {
-            firstCell = await dataGrid.getCellElementExcludingControlColumns(1, 1);
-            expect(await firstCell.getVisibleText()).to.be('debug');
-            await testSubjects.missingOrFail('*logLevelBadgeCell-');
-          });
+          const timestamps = await testSubjects.findAll('exampleRootProfileTimestamp');
+          expect(timestamps).to.have.length(6);
+          expect(await timestamps[0].getVisibleText()).to.be('2024-06-10T16:30:00.000Z');
+          expect(await timestamps[5].getVisibleText()).to.be('2024-06-10T14:00:00.000Z');
         });
       });
 
-      describe('Service Name Cell', () => {
-        it('should render service.name cell', async () => {
+      describe('data source profile', () => {
+        it('should render custom @timestamp but not custom log.level', async () => {
           await common.navigateToActualUrl('discover', undefined, {
             ensureCurrentUrl: false,
           });
-          await header.waitUntilLoadingHasFinished();
+          await dataViews.switchTo('my-example-*');
           await discover.waitUntilSearchingHasFinished();
-          await dataViews.switchToAndValidate('my-example-logs,logstash*');
-          await queryBar.setQuery('service.name:*');
-          await queryBar.submitQuery();
-          await header.waitUntilLoadingHasFinished();
-          await discover.waitUntilSearchingHasFinished();
-          await unifiedFieldList.clickFieldListItemAdd('service.name');
-          await header.waitUntilLoadingHasFinished();
-          await discover.waitUntilSearchingHasFinished();
-
-          let firstCell: WebElementWrapper;
-          let lastCell: WebElementWrapper;
-
-          await retry.try(async () => {
-            firstCell = await dataGrid.getCellElementExcludingControlColumns(0, 1);
-            lastCell = await dataGrid.getCellElementExcludingControlColumns(2, 1);
-            const firstServiceNameCell = await firstCell.findByTestSubject(
-              'dataTableCellActionsPopover_service.name'
-            );
-            const lastServiceNameCell = await lastCell.findByTestSubject(
-              'dataTableCellActionsPopover_service.name'
-            );
-            expect(await firstServiceNameCell.getVisibleText()).to.be('product');
-            expect(await lastServiceNameCell.getVisibleText()).to.be('accounting');
-          });
+          await unifiedFieldList.clickFieldListItemAdd('@timestamp');
+          await unifiedFieldList.clickFieldListItemAdd('log.level');
+          const timestamps = await testSubjects.findAll('exampleRootProfileTimestamp');
+          expect(timestamps).to.have.length(6);
+          expect(await timestamps[0].getVisibleText()).to.be('2024-06-10T16:30:00.000Z');
+          expect(await timestamps[5].getVisibleText()).to.be('2024-06-10T14:00:00.000Z');
+          const logLevels = await testSubjects.findAll('exampleDataSourceProfileLogLevel', 2500);
+          expect(logLevels).to.have.length(0);
         });
 
-        it("should not render service.name cell if it's not a logs data source", async () => {
+        it('should render custom @timestamp and custom log.level', async () => {
           await common.navigateToActualUrl('discover', undefined, {
             ensureCurrentUrl: false,
           });
-          await header.waitUntilLoadingHasFinished();
+          await dataViews.switchTo('my-example-logs');
           await discover.waitUntilSearchingHasFinished();
-          await dataViews.switchToAndValidate('my-example-*');
-          await queryBar.setQuery('service.name:*');
-          await queryBar.submitQuery();
-          await header.waitUntilLoadingHasFinished();
-          await discover.waitUntilSearchingHasFinished();
-          await unifiedFieldList.clickFieldListItemAdd('service.name');
-          await header.waitUntilLoadingHasFinished();
-          await discover.waitUntilSearchingHasFinished();
-
-          let firstCell: WebElementWrapper;
-          let lastCell: WebElementWrapper;
-
-          await retry.try(async () => {
-            firstCell = await dataGrid.getCellElementExcludingControlColumns(0, 1);
-            lastCell = await dataGrid.getCellElementExcludingControlColumns(2, 1);
-
-            expect(await firstCell.getVisibleText()).to.be('product');
-            expect(await lastCell.getVisibleText()).to.be('accounting');
-            await testSubjects.missingOrFail('dataTableCellActionsPopover_service.name');
-          });
-        });
-      });
-
-      describe('Summary column', () => {
-        it('should render a summary of the log entry replacing the original document', async () => {
-          await common.navigateToActualUrl('discover', undefined, {
-            ensureCurrentUrl: false,
-          });
-          await header.waitUntilLoadingHasFinished();
-          await discover.waitUntilSearchingHasFinished();
-          await dataViews.switchToAndValidate('my-example-logs,logstash*');
-
-          await retry.try(async () => {
-            await testSubjects.existOrFail('discoverDataTableMessageValue');
-          });
-        });
-
-        it('should NOT render the summary column if the source does not match logs', async () => {
-          await common.navigateToActualUrl('discover', undefined, {
-            ensureCurrentUrl: false,
-          });
-          await header.waitUntilLoadingHasFinished();
-          await discover.waitUntilSearchingHasFinished();
-          await dataViews.switchToAndValidate('my-example-*');
-
-          await retry.try(async () => {
-            await testSubjects.missingOrFail('discoverDataTableMessageValue');
-          });
+          await unifiedFieldList.clickFieldListItemAdd('@timestamp');
+          await unifiedFieldList.clickFieldListItemAdd('log.level');
+          const timestamps = await testSubjects.findAll('exampleRootProfileTimestamp');
+          expect(timestamps).to.have.length(3);
+          expect(await timestamps[0].getVisibleText()).to.be('2024-06-10T16:00:00.000Z');
+          expect(await timestamps[2].getVisibleText()).to.be('2024-06-10T14:00:00.000Z');
+          const logLevels = await testSubjects.findAll('exampleDataSourceProfileLogLevel');
+          expect(logLevels).to.have.length(3);
+          expect(await logLevels[0].getVisibleText()).to.be('Debug');
+          expect(await logLevels[2].getVisibleText()).to.be('Info');
         });
       });
     });
