+++ test/functional/apps/discover/context_awareness/index.ts
@@ -36,8 +36,7 @@ export default function ({ getService, getPageObjects, loadTestFile }: FtrProvid
       await timePicker.resetDefaultAbsoluteRangeViaUiSettings();
     });
 
-    loadTestFile(require.resolve('./_root_profile'));
-    loadTestFile(require.resolve('./_data_source_profile'));
+    loadTestFile(require.resolve('./_framework'));
     loadTestFile(require.resolve('./_telemetry'));
     loadTestFile(require.resolve('./extensions/_get_row_indicator_provider'));
     loadTestFile(require.resolve('./extensions/_get_row_additional_leading_controls'));
