--- x-pack/test_serverless/functional/test_suites/common/discover/context_awareness/_root_profile.ts
@@ -1,53 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-import kbnRison from '@kbn/rison';
-import expect from '@kbn/expect';
-import type { FtrProviderContext } from '../../../../ftr_provider_context';
-
-export default function ({ getService, getPageObjects }: FtrProviderContext) {
-  const PageObjects = getPageObjects(['common', 'timePicker', 'discover', 'svlCommonPage']);
-  const testSubjects = getService('testSubjects');
-  const dataViews = getService('dataViews');
-
-  describe('root profile', () => {
-    before(async () => {
-      await PageObjects.svlCommonPage.loginAsAdmin();
-    });
-
-    describe('ES|QL mode', () => {
-      describe('cell renderers', () => {
-        it('should not render custom @timestamp', async () => {
-          const state = kbnRison.encode({
-            dataSource: { type: 'esql' },
-            query: { esql: 'from my-example-* | sort @timestamp desc' },
-          });
-          await PageObjects.common.navigateToActualUrl('discover', `?_a=${state}`, {
-            ensureCurrentUrl: false,
-          });
-          await PageObjects.discover.waitUntilSearchingHasFinished();
-          const timestamps = await testSubjects.findAll('exampleRootProfileTimestamp', 2500);
-          expect(timestamps).to.have.length(0);
-        });
-      });
-    });
-
-    describe('data view mode', () => {
-      describe('cell renderers', () => {
-        it('should not render custom @timestamp', async () => {
-          await PageObjects.common.navigateToActualUrl('discover', undefined, {
-            ensureCurrentUrl: false,
-          });
-          await dataViews.switchTo('my-example-*');
-          await PageObjects.discover.waitUntilSearchingHasFinished();
-          const timestamps = await testSubjects.findAll('exampleRootProfileTimestamp', 2500);
-          expect(timestamps).to.have.length(0);
-        });
-      });
-    });
-  });
-}
