+++ x-pack/test_serverless/functional/test_suites/common/discover/context_awareness/extensions/_get_cell_renderers.ts
@@ -7,345 +7,120 @@
 
 import kbnRison from '@kbn/rison';
 import expect from '@kbn/expect';
-import type { WebElementWrapper } from '@kbn/ftr-common-functional-ui-services';
 import type { FtrProviderContext } from '../../../../../ftr_provider_context';
 
 export default function ({ getService, getPageObjects }: FtrProviderContext) {
-  const PageObjects = getPageObjects([
+  const { common, discover, unifiedFieldList, svlCommonPage } = getPageObjects([
     'common',
     'discover',
     'unifiedFieldList',
     'svlCommonPage',
-    'header',
   ]);
-  const esArchiver = getService('esArchiver');
   const testSubjects = getService('testSubjects');
-  const dataGrid = getService('dataGrid');
   const dataViews = getService('dataViews');
-  const queryBar = getService('queryBar');
-  const browser = getService('browser');
-  const retry = getService('retry');
 
   describe('extension getCellRenderers', () => {
     before(async () => {
-      await PageObjects.svlCommonPage.loginAsAdmin();
-      await esArchiver.loadIfNeeded('test/functional/fixtures/es_archiver/logstash_functional');
-    });
-
-    after(async () => {
-      await esArchiver.unload('test/functional/fixtures/es_archiver/logstash_functional');
+      await svlCommonPage.loginAsAdmin();
     });
 
     describe('ES|QL mode', () => {
-      describe('Log Level Badge Cell', () => {
-        it('should render log.level badge cell', async () => {
+      describe('root profile', () => {
+        it('should not render custom @timestamp', async () => {
           const state = kbnRison.encode({
             dataSource: { type: 'esql' },
-            query: {
-              esql: 'from my-example-logs,logstash* | sort @timestamp desc | where `log.level` is not null',
-            },
+            query: { esql: 'from my-example-* | sort @timestamp desc' },
           });
-          await PageObjects.common.navigateToActualUrl('discover', `?_a=${state}`, {
+          await common.navigateToActualUrl('discover', `?_a=${state}`, {
             ensureCurrentUrl: false,
           });
-          await PageObjects.header.waitUntilLoadingHasFinished();
-          await PageObjects.discover.waitUntilSearchingHasFinished();
-          await PageObjects.unifiedFieldList.clickFieldListItemAdd('log.level');
-          await PageObjects.header.waitUntilLoadingHasFinished();
-          await PageObjects.discover.waitUntilSearchingHasFinished();
-
-          const firstCell = await dataGrid.getCellElementExcludingControlColumns(0, 0);
-          const logLevelBadge = await firstCell.findByTestSubject('*logLevelBadgeCell-');
-          expect(await logLevelBadge.getVisibleText()).to.be('debug');
-          expect(await logLevelBadge.getComputedStyle('background-color')).to.be(
-            'rgba(190, 207, 227, 1)'
-          );
-        });
-
-        it("should not render log.level badge cell if it's not a logs data source", async () => {
-          const state = kbnRison.encode({
-            dataSource: { type: 'esql' },
-            query: {
-              esql: 'from my-example* | sort @timestamp desc | where `log.level` is not null',
-            },
-          });
-          await PageObjects.common.navigateToActualUrl('discover', `?_a=${state}`, {
-            ensureCurrentUrl: false,
-          });
-          await PageObjects.header.waitUntilLoadingHasFinished();
-          await PageObjects.discover.waitUntilSearchingHasFinished();
-          await PageObjects.unifiedFieldList.clickFieldListItemAdd('log.level');
-          await PageObjects.header.waitUntilLoadingHasFinished();
-          await PageObjects.discover.waitUntilSearchingHasFinished();
-
-          await retry.try(async () => {
-            const firstCell = await dataGrid.getCellElementExcludingControlColumns(0, 0);
-            expect(await firstCell.getVisibleText()).to.be('debug');
-            await testSubjects.missingOrFail('*logLevelBadgeCell-');
-          });
+          await discover.waitUntilSearchingHasFinished();
+          const timestamps = await testSubjects.findAll('exampleRootProfileTimestamp', 2500);
+          expect(timestamps).to.have.length(0);
         });
       });
 
-      describe('Service Name Cell', () => {
-        it('should render service.name cell', async () => {
+      describe('data source profile', () => {
+        it('should not render custom @timestamp or log.level', async () => {
           const state = kbnRison.encode({
             dataSource: { type: 'esql' },
-            query: {
-              esql: 'from my-example-logs,logstash* | sort @timestamp desc | where `service.name` is not null',
-            },
+            query: { esql: 'from my-example-* | sort @timestamp desc' },
           });
-          await PageObjects.common.navigateToActualUrl('discover', `?_a=${state}`, {
+          await common.navigateToActualUrl('discover', `?_a=${state}`, {
             ensureCurrentUrl: false,
           });
-          await PageObjects.header.waitUntilLoadingHasFinished();
-          await PageObjects.discover.waitUntilSearchingHasFinished();
-          await PageObjects.unifiedFieldList.clickFieldListItemAdd('service.name');
-          await PageObjects.header.waitUntilLoadingHasFinished();
-          await PageObjects.discover.waitUntilSearchingHasFinished();
-
-          const firstCell = await dataGrid.getCellElementExcludingControlColumns(0, 0);
-          const lastCell = await dataGrid.getCellElementExcludingControlColumns(2, 0);
-          const firstServiceNameCell = await firstCell.findByTestSubject(
-            'dataTableCellActionsPopover_service.name'
-          );
-          const lastServiceNameCell = await lastCell.findByTestSubject(
-            'dataTableCellActionsPopover_service.name'
-          );
-          expect(await firstServiceNameCell.getVisibleText()).to.be('product');
-          expect(await lastServiceNameCell.getVisibleText()).to.be('accounting');
+          await discover.waitUntilSearchingHasFinished();
+          await unifiedFieldList.clickFieldListItemAdd('@timestamp');
+          await unifiedFieldList.clickFieldListItemAdd('log.level');
+          const timestamps = await testSubjects.findAll('exampleRootProfileTimestamp', 2500);
+          expect(timestamps).to.have.length(0);
+          const logLevels = await testSubjects.findAll('exampleDataSourceProfileLogLevel', 2500);
+          expect(logLevels).to.have.length(0);
         });
 
-        it("should not render service.name cell if it's not a logs data source", async () => {
+        it('should not render custom @timestamp but should render custom log.level', async () => {
           const state = kbnRison.encode({
             dataSource: { type: 'esql' },
-            query: {
-              esql: 'from my-example* | sort @timestamp desc | where `service.name` is not null',
-            },
+            query: { esql: 'from my-example-logs | sort @timestamp desc' },
           });
-          await PageObjects.common.navigateToActualUrl('discover', `?_a=${state}`, {
+          await common.navigateToActualUrl('discover', `?_a=${state}`, {
             ensureCurrentUrl: false,
           });
-          await PageObjects.header.waitUntilLoadingHasFinished();
-          await PageObjects.discover.waitUntilSearchingHasFinished();
-          await PageObjects.unifiedFieldList.clickFieldListItemAdd('service.name');
-          await PageObjects.header.waitUntilLoadingHasFinished();
-          await PageObjects.discover.waitUntilSearchingHasFinished();
-
-          await retry.try(async () => {
-            const firstCell = await dataGrid.getCellElementExcludingControlColumns(0, 0);
-            expect(await firstCell.getVisibleText()).to.be('product');
-            await testSubjects.missingOrFail('dataTableCellActionsPopover_service.name');
-          });
-        });
-      });
-
-      describe('Summary column', () => {
-        it('should render a summary of the log entry replacing the original document', async () => {
-          const state = kbnRison.encode({
-            dataSource: { type: 'esql' },
-            query: {
-              esql: 'from my-example-logs,logstash* | sort @timestamp desc | where `message` is not null',
-            },
-          });
-          await PageObjects.common.navigateToActualUrl('discover', `?_a=${state}`, {
-            ensureCurrentUrl: false,
-          });
-          await PageObjects.header.waitUntilLoadingHasFinished();
-          await PageObjects.discover.waitUntilSearchingHasFinished();
-
-          await testSubjects.existOrFail('discoverDataTableMessageValue');
-        });
-
-        it('should NOT render the summary column if the source does not match logs', async () => {
-          const state = kbnRison.encode({
-            dataSource: { type: 'esql' },
-            query: {
-              esql: 'from my-example-*',
-            },
-          });
-          await PageObjects.common.navigateToActualUrl('discover', `?_a=${state}`, {
-            ensureCurrentUrl: false,
-          });
-          await PageObjects.header.waitUntilLoadingHasFinished();
-          await PageObjects.discover.waitUntilSearchingHasFinished();
-
-          await testSubjects.missingOrFail('discoverDataTableMessageValue');
+          await discover.waitUntilSearchingHasFinished();
+          await unifiedFieldList.clickFieldListItemAdd('@timestamp');
+          await unifiedFieldList.clickFieldListItemAdd('log.level');
+          const timestamps = await testSubjects.findAll('exampleRootProfileTimestamp', 2500);
+          expect(timestamps).to.have.length(0);
+          const logLevels = await testSubjects.findAll('exampleDataSourceProfileLogLevel');
+          expect(logLevels).to.have.length(3);
+          expect(await logLevels[0].getVisibleText()).to.be('Debug');
+          expect(await logLevels[2].getVisibleText()).to.be('Info');
         });
       });
     });
 
     describe('data view mode', () => {
-      describe('Log Level Badge Cell', () => {
-        it('should render log.level badge cell', async () => {
-          await PageObjects.common.navigateToActualUrl('discover', undefined, {
+      describe('root profile', () => {
+        it('should not render custom @timestamp', async () => {
+          await common.navigateToActualUrl('discover', undefined, {
             ensureCurrentUrl: false,
           });
-          await PageObjects.header.waitUntilLoadingHasFinished();
-          await PageObjects.discover.waitUntilSearchingHasFinished();
-          await dataViews.switchToAndValidate('my-example-logs,logstash*');
-          await queryBar.setQuery('log.level:*');
-          await queryBar.submitQuery();
-          await PageObjects.header.waitUntilLoadingHasFinished();
-          await PageObjects.discover.waitUntilSearchingHasFinished();
-          await PageObjects.unifiedFieldList.clickFieldListItemAdd('log.level');
-          await PageObjects.header.waitUntilLoadingHasFinished();
-          await PageObjects.discover.waitUntilSearchingHasFinished();
-
-          let firstCell: WebElementWrapper;
-          let logLevelBadge: WebElementWrapper;
-
-          await retry.try(async () => {
-            firstCell = await dataGrid.getCellElementExcludingControlColumns(0, 1);
-            logLevelBadge = await firstCell.findByTestSubject('*logLevelBadgeCell-');
-            expect(await logLevelBadge.getVisibleText()).to.be('debug');
-            expect(await logLevelBadge.getComputedStyle('background-color')).to.be(
-              'rgba(190, 207, 227, 1)'
-            );
-          });
-
-          // check Surrounding docs page
-          await dataGrid.clickRowToggle();
-          const [, surroundingActionEl] = await dataGrid.getRowActions();
-          await surroundingActionEl.click();
-          await PageObjects.header.waitUntilLoadingHasFinished();
-          await browser.refresh();
-          await PageObjects.header.waitUntilLoadingHasFinished();
-
-          await retry.try(async () => {
-            firstCell = await dataGrid.getCellElementExcludingControlColumns(0, 1);
-            logLevelBadge = await firstCell.findByTestSubject('*logLevelBadgeCell-');
-            expect(await logLevelBadge.getVisibleText()).to.be('debug');
-            expect(await logLevelBadge.getComputedStyle('background-color')).to.be(
-              'rgba(190, 207, 227, 1)'
-            );
-          });
-        });
-
-        it("should not render log.level badge cell if it's not a logs data source", async () => {
-          await PageObjects.common.navigateToActualUrl('discover', undefined, {
-            ensureCurrentUrl: false,
-          });
-          await PageObjects.header.waitUntilLoadingHasFinished();
-          await PageObjects.discover.waitUntilSearchingHasFinished();
-          await dataViews.switchToAndValidate('my-example-*');
-          await queryBar.setQuery('log.level:*');
-          await queryBar.submitQuery();
-          await PageObjects.header.waitUntilLoadingHasFinished();
-          await PageObjects.discover.waitUntilSearchingHasFinished();
-          await PageObjects.unifiedFieldList.clickFieldListItemAdd('log.level');
-          await PageObjects.header.waitUntilLoadingHasFinished();
-          await PageObjects.discover.waitUntilSearchingHasFinished();
-
-          let firstCell: WebElementWrapper;
-
-          await retry.try(async () => {
-            firstCell = await dataGrid.getCellElementExcludingControlColumns(0, 1);
-            expect(await firstCell.getVisibleText()).to.be('debug');
-            await testSubjects.missingOrFail('*logLevelBadgeCell-');
-          });
-
-          // check Surrounding docs page
-          await dataGrid.clickRowToggle();
-          const [, surroundingActionEl] = await dataGrid.getRowActions();
-          await surroundingActionEl.click();
-          await PageObjects.header.waitUntilLoadingHasFinished();
-          await browser.refresh();
-          await PageObjects.header.waitUntilLoadingHasFinished();
-
-          await retry.try(async () => {
-            firstCell = await dataGrid.getCellElementExcludingControlColumns(1, 1);
-            expect(await firstCell.getVisibleText()).to.be('debug');
-            await testSubjects.missingOrFail('*logLevelBadgeCell-');
-          });
-        });
-      });
-      describe('Service Name Cell', () => {
-        it('should render service.name cell', async () => {
-          await PageObjects.common.navigateToActualUrl('discover', undefined, {
-            ensureCurrentUrl: false,
-          });
-          await PageObjects.header.waitUntilLoadingHasFinished();
-          await PageObjects.discover.waitUntilSearchingHasFinished();
-          await dataViews.switchToAndValidate('my-example-logs,logstash*');
-          await queryBar.setQuery('service.name:*');
-          await queryBar.submitQuery();
-          await PageObjects.header.waitUntilLoadingHasFinished();
-          await PageObjects.discover.waitUntilSearchingHasFinished();
-          await PageObjects.unifiedFieldList.clickFieldListItemAdd('service.name');
-          await PageObjects.header.waitUntilLoadingHasFinished();
-          await PageObjects.discover.waitUntilSearchingHasFinished();
-
-          let firstCell: WebElementWrapper;
-          let lastCell: WebElementWrapper;
-
-          await retry.try(async () => {
-            firstCell = await dataGrid.getCellElementExcludingControlColumns(0, 1);
-            lastCell = await dataGrid.getCellElementExcludingControlColumns(2, 1);
-            const firstServiceNameCell = await firstCell.findByTestSubject(
-              'dataTableCellActionsPopover_service.name'
-            );
-            const lastServiceNameCell = await lastCell.findByTestSubject(
-              'dataTableCellActionsPopover_service.name'
-            );
-            expect(await firstServiceNameCell.getVisibleText()).to.be('product');
-            expect(await lastServiceNameCell.getVisibleText()).to.be('accounting');
-          });
-        });
-
-        it("should not render service.name cell if it's not a logs data source", async () => {
-          await PageObjects.common.navigateToActualUrl('discover', undefined, {
-            ensureCurrentUrl: false,
-          });
-          await PageObjects.header.waitUntilLoadingHasFinished();
-          await PageObjects.discover.waitUntilSearchingHasFinished();
-          await dataViews.switchToAndValidate('my-example-*');
-          await queryBar.setQuery('service.name:*');
-          await queryBar.submitQuery();
-          await PageObjects.header.waitUntilLoadingHasFinished();
-          await PageObjects.discover.waitUntilSearchingHasFinished();
-          await PageObjects.unifiedFieldList.clickFieldListItemAdd('service.name');
-          await PageObjects.header.waitUntilLoadingHasFinished();
-          await PageObjects.discover.waitUntilSearchingHasFinished();
-
-          let firstCell: WebElementWrapper;
-          let lastCell: WebElementWrapper;
-
-          await retry.try(async () => {
-            firstCell = await dataGrid.getCellElementExcludingControlColumns(0, 1);
-            lastCell = await dataGrid.getCellElementExcludingControlColumns(2, 1);
-
-            expect(await firstCell.getVisibleText()).to.be('product');
-            expect(await lastCell.getVisibleText()).to.be('accounting');
-            await testSubjects.missingOrFail('dataTableCellActionsPopover_service.name');
-          });
+          await dataViews.switchTo('my-example-*');
+          await discover.waitUntilSearchingHasFinished();
+          const timestamps = await testSubjects.findAll('exampleRootProfileTimestamp', 2500);
+          expect(timestamps).to.have.length(0);
         });
       });
 
-      describe('Summary column', () => {
-        it('should render a summary of the log entry replacing the original document', async () => {
-          await PageObjects.common.navigateToActualUrl('discover', undefined, {
+      describe('data source profile', () => {
+        it('should not render custom @timestamp or log.level', async () => {
+          await common.navigateToActualUrl('discover', undefined, {
             ensureCurrentUrl: false,
           });
-          await PageObjects.header.waitUntilLoadingHasFinished();
-          await PageObjects.discover.waitUntilSearchingHasFinished();
-          await dataViews.switchToAndValidate('my-example-logs,logstash*');
-
-          await retry.try(async () => {
-            await testSubjects.existOrFail('discoverDataTableMessageValue');
-          });
+          await dataViews.switchTo('my-example-*');
+          await discover.waitUntilSearchingHasFinished();
+          await unifiedFieldList.clickFieldListItemAdd('@timestamp');
+          await unifiedFieldList.clickFieldListItemAdd('log.level');
+          const timestamps = await testSubjects.findAll('exampleRootProfileTimestamp', 2500);
+          expect(timestamps).to.have.length(0);
+          const logLevels = await testSubjects.findAll('exampleDataSourceProfileLogLevel', 2500);
+          expect(logLevels).to.have.length(0);
         });
 
-        it('should NOT render the summary column if the source does not match logs', async () => {
-          await PageObjects.common.navigateToActualUrl('discover', undefined, {
+        it('should not render custom @timestamp but should render custom log.level', async () => {
+          await common.navigateToActualUrl('discover', undefined, {
             ensureCurrentUrl: false,
           });
-          await PageObjects.header.waitUntilLoadingHasFinished();
-          await PageObjects.discover.waitUntilSearchingHasFinished();
-          await dataViews.switchToAndValidate('my-example-*');
-
-          await retry.try(async () => {
-            await testSubjects.missingOrFail('discoverDataTableMessageValue');
-          });
+          await dataViews.switchTo('my-example-logs');
+          await discover.waitUntilSearchingHasFinished();
+          await unifiedFieldList.clickFieldListItemAdd('@timestamp');
+          await unifiedFieldList.clickFieldListItemAdd('log.level');
+          const timestamps = await testSubjects.findAll('exampleRootProfileTimestamp', 2500);
+          expect(timestamps).to.have.length(0);
+          const logLevels = await testSubjects.findAll('exampleDataSourceProfileLogLevel');
+          expect(logLevels).to.have.length(3);
+          expect(await logLevels[0].getVisibleText()).to.be('Debug');
+          expect(await logLevels[2].getVisibleText()).to.be('Info');
         });
       });
     });
