+++ x-pack/test_serverless/functional/test_suites/common/discover/context_awareness/index.ts
@@ -35,8 +35,7 @@ export default function ({ getService, getPageObjects, loadTestFile }: FtrProvid
       await PageObjects.timePicker.resetDefaultAbsoluteRangeViaUiSettings();
     });
 
-    loadTestFile(require.resolve('./_root_profile'));
-    loadTestFile(require.resolve('./_data_source_profile'));
+    loadTestFile(require.resolve('./_framework'));
     loadTestFile(require.resolve('./extensions/_get_row_indicator_provider'));
     loadTestFile(require.resolve('./extensions/_get_row_additional_leading_controls'));
     loadTestFile(require.resolve('./extensions/_get_doc_viewer'));
